Detect Integer literal values at the sql parser
Introduce an integer literal object which will be used instead of a
long literal if the value fits into the integer range.
Queries like ``select 1`` will now return ``integer`` instead of ``long``.

This allows us to later remove the special literal downcast logic on
function matching while keeping the existing behaviour of
``int_col + 1 -> integer``.

+++ server/src/test/java/io/crate/protocols/postgres/PostgresWireProtocolTest.java
@@ -226,7 +226,7 @@ public class PostgresWireProtocolTest extends CrateDummyClusterServiceUnitTest {
             ClientMessages.sendParseMessage(buffer,
                 "S1",
                 "select ? in (1, 2, 3)",
-                new int[] { PGTypes.get(DataTypes.LONG).oid() });
+                new int[] { PGTypes.get(DataTypes.INTEGER).oid() });
                 ClientMessages.sendBindMessage(buffer,
                     "P1",
                     "S1",
@@ -302,7 +302,7 @@ public class PostgresWireProtocolTest extends CrateDummyClusterServiceUnitTest {
                 assertThat(response.readByte(), is((byte) 't'));
                 assertThat(response.readInt(), is(10));
                 assertThat(response.readShort(), is((short) 1));
-                assertThat(response.readInt(), is(PGTypes.get(DataTypes.LONG).oid()));
+                assertThat(response.readInt(), is(PGTypes.get(DataTypes.INTEGER).oid()));
             } finally {
                 response.release();
             }
